Extract AgERA5 data stored in a local hardrive

Extract data from AgERA5 data files previously downloaded from the Copernicus Climate Data Store. These functions use package 'terra' to read *.nc files and extract the requested data for a given location and dates. If dates is one value it extracts a single observation for the specified variable and location. If dates is a character vector of length == 2, it will extract a time series of the specified variable and location, where the first dates value is the start date and the second the end date.

Usage

ag5_extract(coords, ..., path)

# S3 method for numeric
ag5_extract(
  coords,
  dates,
  variable,
  statistic = NULL,
  time = NULL,
  celsius = FALSE,
  ...,
  path
)

# S3 method for data.frame
ag5_extract(
  coords,
  lon = "lon",
  lat = "lat",
  start_date = "start_date",
  end_date = "end_date",
  variable,
  statistic = NULL,
  time = NULL,
  celsius = FALSE,
  ...,
  path
)

Arguments

coords

numeric vector of length = 2 of the form (lon, lat), or a data.frame with required columns

...

Other parameters

path

character The path for the folder containing the AgERA5 files

dates

character The dates for extracting the specified variable, a vector of length 1 extracts a single date, while a vector of length 2 indicates the start and end dates. or the column name in the case of data.frame

variable

character The AgERA5 variable to extract, see details for available options

statistic

character Only for some variables, see details for valid options

time

Only for variable Relative-Humidity-2m, see details for valid options

celsius

logical Only for variables "Temperature-Air-2m" and "2m_dewpoint_temperature". If TRUE the values are converted from Kelvin to Celsius. Default is FALSE

lon

character Column name of longitude values in the case of data.frame

lat

character Column name of latitude values in the case of data.frame

start_date

character Column name of start_date values in the case that coords is a data.frame

end_date

character Column name of end_date values in the case that coords is a data.frame

Valid variable values

"cloud_cover"

"liquid_precipitation_duration_fraction"

"snow_thickness_lwe"

"Solar-Radiation-Flux"

"Temperature-Air-2m"

"2m_dewpoint_temperature"

"Precipitation-Flux"

"solid_precipitation_duration_fraction"

"snow_thickness"

"vapour_pressure"

"10m_wind_speed"

"Relative-Humidity-2m"

Valid statistics for variable "Temperature-Air-2m"

Max-24h

Mean-24h

24_hour_minimum

Max-Day-Time

Mean-Day-Time

Mean-Night-Time

Min-Night-Time

Variables that require statistic

For the following variables, only "24_hour_mean" statistic is available, but should be explicitly indicated.

cloud_cover

snow_thickness_lwe

2m_dewpoint_temperature

snow_thickness

vapour_pressure

10m_wind_speed

Valid time values for variable "Relative-Humidity-2m"

06h

09h

12h

15h

18h

References

Temperature conversion is made accordingly to: Preston-Thomas, H. (1990). The International Temperature Scale of 1990 (ITS-90). Metrologia, 27(1), 3-10. doi:10.1088/0026-1394/27/1/002

Examples

if (FALSE) {
temp <- ag5_extract(coords = c(lon = 35.72636, lat = -2.197162),
                      dates = "1991-04-22",
                      variable = "Temperature-Air-2m",
                      statistic = "Max-Day-Time",
                      path = "C:/temperature_data/")
}